Move tests into feature namespace
Just to be consistent
