+++ phpunit.xml.dist
@@ -9,8 +9,8 @@
          processIsolation="false"
          stopOnFailure="false">
     <testsuites>
-        <testsuite name="Hyde">
-            <directory suffix="Test.php">./tests</directory>
+        <testsuite name="FeatureHyde">
+            <directory suffix="Test.php">./tests/Hyde/Feature</directory>
         </testsuite>
         <testsuite name="FeatureFramework">
             <directory suffix="Test.php">./packages/framework/tests/Feature</directory>
