+++ tests/Hyde/Feature/DefaultContentTest.php
@@ -1,10 +1,9 @@
 <?php
 
-namespace Hyde\Testing\Hyde;
+namespace Hyde\Testing\Hyde\Feature;
 
 use Hyde\Framework\Hyde;
 use Hyde\Testing\TestCase;
-use Illuminate\Support\Facades\File;
 
 class DefaultContentTest extends TestCase
 {
