+++ tests/Hyde/Feature/HydeCLITest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Hyde\Testing\Hyde;
+namespace Hyde\Testing\Hyde\Feature;
 
 use Hyde\Testing\TestCase;
